Save flight in cloud function in transaction with techlog entry
Before, the flight was saved in the client and the techlog entry
was saved in a cloud function and there were cases where the flight
(with techlog entry status and description) could be saved but no
actual techlog entry was created. This should never happen.
