+++ functions/aircraft/addTechlogEntry.function.js
@@ -1,7 +1,7 @@
 const functions = require('firebase-functions')
 const admin = require('firebase-admin')
 const getMemberByUid = require('../utils/getMemberByUid')
-const addAttachments = require('./addAttachments')
+const addEntry = require('./addTechlogEntry')
 
 // Prevent firebase from initializing twice
 try {
@@ -9,85 +9,27 @@ try {
   // eslint-disable-next-line no-empty
 } catch (e) {}
 
-const ALLOWED_USER_STATUS = [
-  'for_information_only',
-  'defect_aog',
-  'defect_unknown'
-]
-
 const db = admin.firestore()
 const bucket = admin.storage().bucket()
 
 const addTechlogEntry = functions.https.onCall(async (data, context) => {
   const { organizationId, aircraftId, entry } = data
-
-  const member = await getMemberByUid(db, organizationId, context.auth.uid)
-
-  if (entry.initialStatus !== entry.currentStatus) {
-    throw new Error(
-      `initialStatus '${entry.initialStatus}' is not equal to currentStatus '${entry.currentStatus}'`
-    )
-  }
-
-  if (
-    !member.get('roles').includes('techlogmanager') &&
-    !ALLOWED_USER_STATUS.includes(entry.initialStatus)
-  ) {
-    throw new Error(
-      `Status '${entry.initialStatus}' can only be set by techlogmanager`
-    )
-  }
-
-  const author = {
-    firstname: member.get('firstname'),
-    lastname: member.get('lastname'),
-    nr: member.get('nr') || null,
-    member: member.ref,
-    id: member.id
-  }
-
-  entry.timestamp = new Date()
-  entry.deleted = false
-  entry.author = author
-  if (entry.closed === true) {
-    entry.closedTimestamp = admin.firestore.FieldValue.serverTimestamp()
-    entry.closedBy = author
-  }
-
-  await db.runTransaction(async t => {
-    const aircraftRef = db
-      .collection('organizations')
-      .doc(organizationId)
-      .collection('aircrafts')
-      .doc(aircraftId)
-    const newEntryRef = aircraftRef.collection('techlog').doc()
-
-    const aircraftDoc = await t.get(aircraftRef)
-
-    if (aircraftDoc.exists !== true) {
-      throw new Error(
-        `Aircraft ${aircraftId} in organization ${organizationId} does not exist`
-      )
-    }
-
-    const techlogEntriesCount = aircraftDoc.get('counters.techlogEntries') || 0
-    const newCount = techlogEntriesCount + 1
-
-    entry.number = newCount
-    entry.attachments = await addAttachments(
-      bucket,
+  const member = await getMemberByUid(db, data.organizationId, context.auth.uid)
+  const aircraftRef = db
+    .collection('organizations')
+    .doc(organizationId)
+    .collection('aircrafts')
+    .doc(aircraftId)
+  await db.runTransaction(async transaction => {
+    await addEntry(
       organizationId,
       aircraftId,
-      newEntryRef.id,
-      null,
-      entry.attachments
+      entry,
+      member,
+      aircraftRef,
+      bucket,
+      transaction
     )
-
-    await t.set(newEntryRef, entry)
-
-    await t.update(aircraftRef, {
-      'counters.techlogEntries': newCount
-    })
   })
 })
