+++ functions/aircraft/mergeDateAndTime.js
@@ -1,0 +1,9 @@
+const moment = require('moment-timezone')
+
+const mergeDateAndTime = (date, time, timezone) => {
+  const dateTimeFormat = 'YYYY-MM-DD HH:mm'
+  const timeString = moment.tz(time, dateTimeFormat, timezone).format('HH:mm')
+  return moment.tz(date + ' ' + timeString, dateTimeFormat, timezone).toDate()
+}
+
+module.exports = mergeDateAndTime
