+++ src/routes/organizations/routes/aircraft/module/sagas.js
@@ -12,19 +12,12 @@ import {
   getDoc,
   addDoc,
   updateDoc,
-  runTransaction,
   serverTimestamp
 } from '../../../../../util/firestoreUtils'
 import getLastFlight from './util/getLastFlight'
-import { validateSync, validateAsync } from './util/validateFlight'
-import { getCounters } from './util/counters'
 import { fetchAerodromes, fetchAircrafts } from '../../../module/actions'
 import { isClosed } from '../../../../../util/techlogStatus'
-import {
-  getCurrentMember,
-  getCurrentMemberObject,
-  getMemberObject
-} from '../../../util/members'
+import { getCurrentMember, getCurrentMemberObject } from '../../../util/members'
 
 export const flightPageName = (aircraftId, page, showDeleted) =>
   `${showDeleted ? 'flights-all' : 'flights'}-${aircraftId}-${page}`
@@ -135,22 +128,6 @@ export function* fetchFlights() {
   )
 }
 
-export function* getAerodrome(organizationId, aerodromeId) {
-  let aerodrome = yield call(getDoc, ['aerodromes', aerodromeId])
-  if (aerodrome.exists !== true) {
-    aerodrome = yield call(getDoc, [
-      'organizations',
-      organizationId,
-      'aerodromes',
-      aerodromeId
-    ])
-  }
-  if (aerodrome.exists !== true) {
-    throw new Error(`Aeorodrome not found for id ${aerodromeId}`)
-  }
-  return aerodrome
-}
-
 export function* getFlight(organizationId, aircraftId, flightId) {
   return yield call(getDoc, [
     'organizations',
@@ -186,25 +163,11 @@ export async function getDestinationAerodrome(flight) {
   return null
 }
 
-export const mergeDateAndTime = (date, time, timezone) => {
-  const dateTimeFormat = 'YYYY-MM-DD HH:mm'
-  const timeString = moment.tz(time, dateTimeFormat, timezone).format('HH:mm')
-  return moment.tz(date + ' ' + timeString, dateTimeFormat, timezone).toDate()
-}
-
 export const extractDate = (timestamp, timezone) =>
   moment(timestamp.toDate()).tz(timezone).format('YYYY-MM-DD')
 
 export const getTimeForPicker = (timestamp, timezone) =>
   moment(timestamp.toDate()).tz(timezone).format('YYYY-MM-DD HH:mm')
-
-export const aerodromeObject = aeodromeDocument => ({
-  name: aeodromeDocument.get('name') || null,
-  identification: aeodromeDocument.get('identification') || null,
-  timezone: aeodromeDocument.get('timezone') || null,
-  aerodrome: aeodromeDocument.ref,
-  id: aeodromeDocument.id
-})
 
 export function* createFlight({
   payload: { organizationId, aircraftId, data }
@@ -212,170 +175,34 @@ export function* createFlight({
   try {
     yield put(actions.setCreateFlightDialogSubmitting())
 
-    const aircraftSettings = yield select(aircraftSettingsSelector, aircraftId)
-
-    let validationErrors = yield call(
-      validateSync,
-      data,
-      organizationId,
-      aircraftId,
-      aircraftSettings
-    )
-    if (Object.getOwnPropertyNames(validationErrors).length > 0) {
-      yield put(actions.setFlightValidationErrors(validationErrors))
-      return
-    }
-
-    const departureAerodrome = yield call(
-      getAerodrome,
-      organizationId,
-      data.departureAerodrome.value
-    )
-    const destinationAerodrome = yield call(
-      getAerodrome,
-      organizationId,
-      data.destinationAerodrome.value
-    )
-
-    const counters = getCounters(data)
-
-    data = {
+    const dataToSave = {
       ...data,
-      blockOffTime: mergeDateAndTime(
-        data.date,
-        data.blockOffTime,
-        departureAerodrome.get('timezone')
-      ),
-      takeOffTime: mergeDateAndTime(
-        data.date,
-        data.takeOffTime,
-        departureAerodrome.get('timezone')
-      ),
-      landingTime: mergeDateAndTime(
-        data.date,
-        data.landingTime,
-        destinationAerodrome.get('timezone')
-      ),
-      blockOnTime: mergeDateAndTime(
-        data.date,
-        data.blockOnTime,
-        destinationAerodrome.get('timezone')
+      techlogEntryAttachments: yield call(
+        getAttachments,
+        data.techlogEntryAttachments || []
       )
     }
 
-    validationErrors = yield call(
-      validateAsync,
-      data,
+    const result = yield call(callFunction, 'saveFlight', {
       organizationId,
-      aircraftId
-    )
-    if (Object.getOwnPropertyNames(validationErrors).length > 0) {
-      yield put(actions.setFlightValidationErrors(validationErrors))
+      aircraftId,
+      data: dataToSave,
+      techlogEntryClosed:
+        data.techlogEntryStatus && data.techlogEntryStatus.value
+          ? isClosed(data.techlogEntryStatus.value)
+          : null
+    })
+
+    if (
+      result.data &&
+      result.data.validationErrors &&
+      Object.getOwnPropertyNames(result.data.validationErrors).length > 0
+    ) {
+      yield put(actions.setFlightValidationErrors(result.data.validationErrors))
       return
     }
 
-    const owner = yield call(getCurrentMemberObject, organizationId)
-    const pilot = yield call(getMemberObject, organizationId, data.pilot.value)
-    const instructor =
-      data.instructor && data.instructor.value
-        ? yield call(getMemberObject, organizationId, data.instructor.value)
-        : null
-
-    const fuelUplift =
-      typeof data.fuelUplift === 'number' ? data.fuelUplift / 100 : null
-    const fuelType =
-      typeof fuelUplift === 'number' && fuelUplift > 0
-        ? data.fuelType.value
-        : null
-
-    const oilUplift =
-      typeof data.oilUplift === 'number' ? data.oilUplift / 100 : null
-
-    const timestampFieldValue = yield call(serverTimestamp)
-
-    const dataToStore = {
-      deleted: false,
-      version: 1,
-      owner,
-      createTimestamp: timestampFieldValue,
-      nature: typeof data.nature === 'string' ? data.nature : data.nature.value,
-      pilot,
-      instructor,
-      departureAerodrome: aerodromeObject(departureAerodrome),
-      destinationAerodrome: aerodromeObject(destinationAerodrome),
-      blockOffTime: data.blockOffTime,
-      takeOffTime: data.takeOffTime,
-      landingTime: data.landingTime,
-      blockOnTime: data.blockOnTime,
-      counters,
-      landings: data.landings,
-      personsOnBoard: data.personsOnBoard,
-      fuelUplift,
-      fuelType,
-      fuelUnit: 'litre',
-      oilUplift,
-      oilUnit: 'litre',
-      remarks: data.remarks || null,
-      preflightCheck:
-        typeof data.preflightCheck === 'boolean' ? data.preflightCheck : null
-    }
-
-    if (data.troublesObservations) {
-      dataToStore.troublesObservations = data.troublesObservations
-      dataToStore.techlogEntryDescription = data.techlogEntryDescription
-        ? data.techlogEntryDescription.trim()
-        : null
-      if (aircraftSettings.techlogEnabled === true) {
-        dataToStore.techlogEntryStatus = data.techlogEntryStatus
-          ? typeof data.techlogEntryStatus === 'string'
-            ? data.techlogEntryStatus
-            : data.techlogEntryStatus.value
-          : null
-      }
-    }
-
-    const oldFlightDoc = data.id
-      ? yield call(getDoc, [
-          'organizations',
-          organizationId,
-          'aircrafts',
-          aircraftId,
-          'flights',
-          data.id
-        ])
-      : null
-    const newFlightDoc = yield call(addNewFlightDoc, organizationId, aircraftId)
-    yield call(
-      runTransaction,
-      setFlightData,
-      oldFlightDoc,
-      newFlightDoc,
-      dataToStore,
-      owner,
-      timestampFieldValue
-    )
-
-    if (
-      data.troublesObservations === 'troubles' &&
-      !data.id &&
-      aircraftSettings.techlogEnabled === true
-    ) {
-      const entry = {
-        description: data.techlogEntryDescription.trim(),
-        initialStatus: data.techlogEntryStatus.value,
-        currentStatus: data.techlogEntryStatus.value,
-        closed: isClosed(data.techlogEntryStatus.value),
-        flight: newFlightDoc.id,
-        attachments: yield call(
-          getAttachments,
-          data.techlogEntryAttachments || []
-        )
-      }
-      yield call(callFunction, 'addTechlogEntry', {
-        organizationId,
-        aircraftId,
-        entry
-      })
+    if (result.data && result.data.techlogEntryAdded) {
       yield put(fetchAircrafts(organizationId))
       yield call(fetchTechlog)
     }
@@ -386,43 +213,6 @@ export function* createFlight({
     error(`Failed to create flight`, e)
     yield put(actions.createFlightFailure())
   }
-}
-
-export function* addNewFlightDoc(organizationId, aircraftId) {
-  const newDoc = yield call(
-    addDoc,
-    ['organizations', organizationId, 'aircrafts', aircraftId, 'flights'],
-    { deleted: true }
-  )
-  // fetch again with ref
-  return yield call(getDoc, [
-    'organizations',
-    organizationId,
-    'aircrafts',
-    aircraftId,
-    'flights',
-    newDoc.id
-  ])
-}
-
-export const setFlightData = (
-  oldFlightDoc,
-  newFlightDoc,
-  dataToStore,
-  currentMember,
-  timestampFieldValue
-) => async tx => {
-  if (oldFlightDoc) {
-    tx.update(oldFlightDoc.ref, {
-      deleted: true,
-      replacedWith: newFlightDoc.id,
-      deletedBy: currentMember,
-      deleteTimestamp: timestampFieldValue
-    })
-    dataToStore.replaces = oldFlightDoc.id
-    dataToStore.version = oldFlightDoc.get('version') + 1
-  }
-  tx.update(newFlightDoc.ref, dataToStore)
 }
 
 export function* initCreateFlightDialog({
@@ -860,7 +650,7 @@ export function* fetchLatestCrs({ payload: { organizationId, aircraftId } }) {
   }
 }
 
-const getBase64 = file =>
+export const getBase64 = file =>
   new Promise((resolve, reject) => {
     const reader = new FileReader()
     reader.readAsDataURL(file)
