+++ src/util/dates.spec.js
@@ -1,10 +1,4 @@
-import {
-  formatDate,
-  formatTime,
-  addHours,
-  getTimeDiff,
-  getTimeDiffInHundredthsOfHour
-} from './dates'
+import { formatDate, formatTime, addHours, getTimeDiff } from './dates'
 
 describe('util', () => {
   describe('dates', () => {
@@ -51,32 +45,6 @@ describe('util', () => {
         expect(
           addHours('2018-11-20 10:00', 2.1, 'Europe/Zurich', 'Europe/Athens')
         ).toEqual('2018-11-20 13:06')
-      })
-    })
-
-    describe('getTimeDiffInHundredthsOfHour', () => {
-      it('should return the difference between two timestamps in hundredths of an hour', () => {
-        expect(
-          getTimeDiffInHundredthsOfHour('2018-11-20 10:00', '2018-11-20 11:00')
-        ).toEqual(100)
-      })
-
-      it('should round to integers', () => {
-        expect(
-          getTimeDiffInHundredthsOfHour('2018-11-20 10:00', '2018-11-20 10:01')
-        ).toEqual(2)
-      })
-
-      it('should return 0 if timestamps are equal', () => {
-        expect(
-          getTimeDiffInHundredthsOfHour('2018-11-20 10:00', '2018-11-20 10:00')
-        ).toEqual(0)
-      })
-
-      it('should return negative numbers if start timestamp is after end timestamp', () => {
-        expect(
-          getTimeDiffInHundredthsOfHour('2018-11-20 10:00', '2018-11-20 09:30')
-        ).toEqual(-50)
       })
     })
   })
